requirements.txt
# Dépendances pour l'extraction de données (si nécessaire)
# beautifulsoup4==4.13.3
# langdetect==1.0.9
# nltk==3.9.1
# scholarly==1.7.11

# Dépendances pour le modèle (versions compatibles)
numpy==1.20.3  # Version plus compatible avec scikit-learn 1.3.0
pandas==2.0.3  # Version compatible avec numpy 1.19.5
scikit-learn==1.0.2  # Version compatible avec numpy 1.19.5
torch==2.0.0 --index-url https://download.pytorch.org/whl/cpu

# Dépendances pour la base de données
pymongo[srv]==3.12.0

# Dépendances pour les requêtes HTTP
requests==2.32.3

# Version minimale d'Airflow sans extras non nécessaires
# Version d'Airflow qui respecte les contraintes de dépendances
apache-airflow==2.7.0

# Downgrade Flask pour être compatible avec Airflow
flask==2.2.5

# MLflow avec version fixée
mlflow==2.8.0
python-dotenv==1.0.0
